Added watchOS target and some stream fixes

+++ Radio-macOS/Controllers/PlayerViewController.swift
@@ -90,7 +90,7 @@ class PlayerViewController: NSViewController, RadioPlayerDelegate {
         }
     }
     
-    func radioUpdatedTime(currentTime: Int) {
+    func radioUpdatedTime(currentTime: Double) {
         
     }
